MID-8929 Improve processing of fetching existing entry when updating it in AD connector.
Up to now, the entry was fetched by DN, but as there is a case where the entry is renamed and cannot be fetched,
the fetch is switched to GUID-based search.
In addition, the process has been improved so that existing entry is fetched when it is needed to improve performance.