feat: Update SunChangelogSyncStrategy to be configurable (and hence generic)
As part of this commit, the configuration and SunChangelogSyncStrategy have been extended to be more configurable; primarily with the ability to change the attribute names used by the sync strategy. This allows for other directory software that uses a similiar approach to change logs as SunDS does, such as Isode M-Vault, to also work with this sync strategy.
Although the class has been renamed as part of this commit to reflect the more generic nature of the upgraded class, the default configuration options, and SyncStrategy option of "sunChangeLog" have been retained for backwards-compatibility.